chore: release v1.5.66 — Clippy compliance pass across all crates
Resolves all outstanding Clippy warnings workspace-wide:
- #[must_use] annotations on all pure functions in sloc-languages/style helpers
- const fn conversions for deterministic helpers across sloc-languages, sloc-core, sloc-mcp
- f64::from() replaces as f32 casts (cast_precision_loss) throughout style analyzers
- map_or replaces .map(...).unwrap_or(...) idiom across sloc-cli, sloc-core, sloc-report, sloc-languages
- u32::try_from replaces as u32 usize casts (cast_possible_truncation) in count_over / scan_base_metrics
- Let-else replaces match-on-option in scan_indent
- Explicit imports replace wildcard use super::common::* in rust_lang.rs
- McpConfig::from_env() simplified to return Self (infallible, no Result wrapper needed)
- McpResponse::parse_error takes &serde_json::Error instead of owned
- #[allow(clippy::...)] suppression annotations with justification comments
- Doc comments backtick-quote identifiers and env-var names for rustdoc
- dist/oxide-sloc-windows-x64.zip updated with freshly built v1.5.66 binary

+++ crates/sloc-languages/src/style/python.rs
@@ -32,9 +32,9 @@ pub fn analyze(text: &str) -> StyleAnalysis {
 
     let quote_val = if single_q == 0 && double_q == 0 {
         "\u{2014}"
-    } else if double_q as f32 / (single_q + double_q) as f32 >= 0.70 {
+    } else if f64::from(double_q) / f64::from(single_q + double_q) >= 0.70 {
         "Double quotes"
-    } else if single_q as f32 / (single_q + double_q) as f32 >= 0.70 {
+    } else if f64::from(single_q) / f64::from(single_q + double_q) >= 0.70 {
         "Single quotes"
     } else {
         "Mixed"
@@ -71,6 +71,7 @@ fn has_type_hints(trimmed: &str) -> bool {
         && (trimmed.contains(": ") || trimmed.contains("->"))
 }
 
+#[allow(clippy::cast_precision_loss)] // counts bounded well within f32 precision
 fn score_double_quotes(double: u32, single: u32) -> f32 {
     let t = double + single;
     if t == 0 {
